Improvements from code review - some changes still not addressed
