+++ calcite-adapter/src/test/java/software/amazon/documentdb/jdbc/query/DocumentDbQueryMappingServiceDateTimeTest.java
@@ -16,86 +16,37 @@
 
 package software.amazon.documentdb.jdbc.query;
 
-import com.mongodb.client.MongoClient;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import org.bson.BsonArray;
 import org.bson.BsonDateTime;
 import org.bson.BsonDocument;
 import org.bson.BsonInt64;
 import org.bson.BsonString;
 import org.bson.conversions.Bson;
-import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import software.amazon.documentdb.jdbc.DocumentDbConnectionProperties;
 import software.amazon.documentdb.jdbc.calcite.adapter.DocumentDbFilter;
 import software.amazon.documentdb.jdbc.common.test.DocumentDbFlapDoodleExtension;
-import software.amazon.documentdb.jdbc.common.test.DocumentDbFlapDoodleTest;
-import software.amazon.documentdb.jdbc.metadata.DocumentDbDatabaseSchemaMetadata;
-import software.amazon.documentdb.jdbc.persist.SchemaStoreFactory;
-import software.amazon.documentdb.jdbc.persist.SchemaWriter;
 
 import java.sql.SQLException;
 import java.time.Instant;
 import java.util.List;
 
-import static software.amazon.documentdb.jdbc.metadata.DocumentDbDatabaseSchemaMetadata.VERSION_NEW;
-
 @ExtendWith(DocumentDbFlapDoodleExtension.class)
-public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoodleTest {
-    private static final String DATABASE_NAME = "database";
-    private static final String USER = "user";
-    private static final String PASSWORD = "password";
-    private static final String COLLECTION_NAME = "testCollection";
-    private static final String OTHER_COLLECTION_NAME = "otherTestCollection";
+public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbQueryMappingServiceTest {
     private static final String DATE_COLLECTION_NAME = "dateTestCollection";
     private static DocumentDbQueryMappingService queryMapper;
-    private static DocumentDbConnectionProperties connectionProperties;
-    private static MongoClient client;
 
     @BeforeAll
-    @SuppressFBWarnings(value = "HARD_CODE_PASSWORD", justification = "Hardcoded for test purposes only")
-    static void initialize() throws SQLException {
-        // Add a valid users to the local MongoDB instance.
-        connectionProperties = new DocumentDbConnectionProperties();
-        createUser(DATABASE_NAME, USER, PASSWORD);
-        connectionProperties.setUser(USER);
-        connectionProperties.setPassword(PASSWORD);
-        connectionProperties.setDatabase(DATABASE_NAME);
-        connectionProperties.setTlsEnabled("false");
-        connectionProperties.setHostname("localhost:" + getMongoPort());
+    void initialize() throws SQLException {
         final long dateTime = Instant.parse("2020-01-01T00:00:00.00Z").toEpochMilli();
-        final BsonDocument document =
-                BsonDocument.parse(
-                        "{ \"_id\" : \"key\", \"array\" : [ { \"field\" : 1, \"field1\": \"value\" }, { \"field\" : 2, \"field2\" : \"value\" } ]}");
-
-        final BsonDocument otherDocument =
-                BsonDocument.parse(
-                        "{ \"_id\" : \"key1\", \"otherArray\" : [ { \"field\" : 1, \"field3\": \"value\" }, { \"field\" : 2, \"field3\" : \"value\" } ]}");
-        final BsonDocument doc1 = BsonDocument.parse("{\"_id\": 101}");
-        doc1.append("field", new BsonDateTime(dateTime));
-        client = createMongoClient(ADMIN_DATABASE, USER, PASSWORD);
-
-        insertBsonDocuments(
-                COLLECTION_NAME, DATABASE_NAME, new BsonDocument[]{document}, client);
-        insertBsonDocuments(
-                OTHER_COLLECTION_NAME, DATABASE_NAME, new BsonDocument[]{otherDocument}, client);
-        insertBsonDocuments(DATE_COLLECTION_NAME, DATABASE_NAME, new BsonDocument[]{doc1}, client);
-        final DocumentDbDatabaseSchemaMetadata databaseMetadata =
-                DocumentDbDatabaseSchemaMetadata.get(connectionProperties, "id", VERSION_NEW, client);
-        queryMapper = new DocumentDbQueryMappingService(connectionProperties, databaseMetadata, client);
-    }
-
-    @AfterAll
-    static void afterAll() throws Exception {
-        try (SchemaWriter schemaWriter = SchemaStoreFactory.createWriter(connectionProperties, client)) {
-            schemaWriter.remove("id");
-        }
-        client.close();
+        final BsonDocument document = BsonDocument.parse("{\"_id\": 101}");
+        document.append("field", new BsonDateTime(dateTime));
+        insertBsonDocuments(DATE_COLLECTION_NAME, new BsonDocument[]{document});
+        queryMapper = getQueryMappingService();
     }
 
     /**
@@ -115,7 +66,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                                 + "TIMESTAMPADD(SECOND, 5, \"field\"), "
                                 + "TIMESTAMPADD(MICROSECOND, 6, \"field\") "
                                 + "FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         DocumentDbMqlQueryContext result = queryMapper.get(timestampAddQuery);
         Assertions.assertNotNull(result);
         Assertions.assertEquals(DATE_COLLECTION_NAME, result.getCollectionName());
@@ -146,7 +97,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                                 + "SECOND(\"field\"),"
                                 + "QUARTER(\"field\")"
                                 + "FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         result = queryMapper.get(extractQuery);
         Assertions.assertNotNull(result);
         Assertions.assertEquals(DATE_COLLECTION_NAME, result.getCollectionName());
@@ -186,7 +137,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                                 + "TIMESTAMPDIFF(QUARTER, \"field\", \"field\"), "
                                 + "TIMESTAMPDIFF(MONTH, \"field\", \"field\")"
                                 + "FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         result = queryMapper.get(timestampDiffQuery);
         Assertions.assertNotNull(result);
         Assertions.assertEquals(DATE_COLLECTION_NAME, result.getCollectionName());
@@ -220,7 +171,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT CURRENT_TIMESTAMP AS \"cts\""
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext result1 = queryMapper.get(currentTimestampQuery);
         Assertions.assertNotNull(result1);
         Assertions.assertEquals(DATE_COLLECTION_NAME, result1.getCollectionName());
@@ -248,7 +199,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT CURRENT_DATE AS \"cts\""
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext result2 = queryMapper.get(currentDateQuery);
         Assertions.assertNotNull(result2);
         Assertions.assertEquals(DATE_COLLECTION_NAME, result2.getCollectionName());
@@ -276,7 +227,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT CURRENT_TIME AS \"cts\""
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext result3 = queryMapper.get(currentTimeQuery);
         Assertions.assertNotNull(result3);
         Assertions.assertEquals(DATE_COLLECTION_NAME, result3.getCollectionName());
@@ -311,7 +262,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT TIMESTAMPADD(MONTH, 10, \"field\") AS \"cts\""
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         Assertions.assertEquals(String.format("Unable to parse SQL"
                         + " 'SELECT TIMESTAMPADD(MONTH, 10, \"field\") AS \"cts\" FROM \"database\".\"dateTestCollection\"'.%n"
                         + " Reason: 'Conversion between the source type (INTERVAL_MONTH) and the target type (TIMESTAMP) is not supported.'"),
@@ -322,7 +273,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT TIMESTAMPADD(YEAR, 10, \"field\") AS \"cts\""
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         Assertions.assertEquals(String.format("Unable to parse SQL"
                         + " 'SELECT TIMESTAMPADD(YEAR, 10, \"field\") AS \"cts\" FROM \"database\".\"dateTestCollection\"'.%n"
                         + " Reason: 'Conversion between the source type (INTERVAL_YEAR) and the target type (TIMESTAMP) is not supported.'"),
@@ -333,7 +284,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT TIMESTAMPADD(QUARTER, 10, \"field\") AS \"cts\""
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         Assertions.assertEquals(String.format("Unable to parse SQL"
                         + " 'SELECT TIMESTAMPADD(QUARTER, 10, \"field\") AS \"cts\" FROM \"database\".\"dateTestCollection\"'.%n"
                         + " Reason: 'Conversion between the source type (INTERVAL_MONTH) and the target type (TIMESTAMP) is not supported.'"),
@@ -352,7 +303,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT DAYNAME(\"field\") AS \"cts\""
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context = queryMapper.get(dayNameQuery);
         Assertions.assertNotNull(context);
         final List<Bson> operations = context.getAggregateOperations();
@@ -373,7 +324,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT DAYNAME(NULL) AS \"cts\""
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context2 = queryMapper.get(dayNameQuery2);
         Assertions.assertNotNull(context2);
         final List<Bson> operations2 = context2.getAggregateOperations();
@@ -402,7 +353,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT MONTHNAME(\"field\") AS \"cts\""
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context = queryMapper.get(dayNameQuery);
         Assertions.assertNotNull(context);
         final List<Bson> operations = context.getAggregateOperations();
@@ -428,7 +379,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT MONTHNAME(NULL) AS \"cts\""
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context2 = queryMapper.get(dayNameQuery2);
         Assertions.assertNotNull(context2);
         final List<Bson> operations2 = context2.getAggregateOperations();
@@ -465,7 +416,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                                 + " FLOOR(\"field\" TO SECOND),"
                                 + " FLOOR(\"field\" TO MILLISECOND)"
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context = queryMapper.get(floorDayQuery);
         Assertions.assertNotNull(context);
         final List<Bson> operations = context.getAggregateOperations();
@@ -509,7 +460,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                         "SELECT"
                                 + " FLOOR(\"field\" TO WEEK)"
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context1 = queryMapper.get(floorDayQuery1);
         Assertions.assertNotNull(context1);
         final List<Bson> operations1 = context1.getAggregateOperations();
@@ -534,7 +485,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                                 + "   TIMESTAMP '1970-01-05', \"field\"),%n"
                                 + " TIMESTAMP '1970-01-05')%n"
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context2 = queryMapper.get(floorDayQuery2);
         Assertions.assertNotNull(context2);
         final List<Bson> operations2 = context2.getAggregateOperations();
@@ -597,7 +548,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT FLOOR(\"field\" TO QUARTER)"
                                 + " FROM \"%s\".\"%s\"",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context3 = queryMapper.get(floorDayQuery3);
         Assertions.assertNotNull(context3);
         final List<Bson> operations3 = context3.getAggregateOperations();
@@ -624,7 +575,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
         final String dayNameQuery =
                 String.format(
                         "SELECT * FROM \"%s\".\"%s\" WHERE MONTHNAME(\"field\") = 'February'",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context = queryMapper.get(dayNameQuery);
         Assertions.assertNotNull(context);
         final List<Bson> operations = context.getAggregateOperations();
@@ -663,7 +614,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
         final String dayNameQuery =
                 String.format(
                         "SELECT * FROM \"%s\".\"%s\" WHERE DAYNAME(\"field\") = 'Tuesday'",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context = queryMapper.get(dayNameQuery);
         Assertions.assertNotNull(context);
         final List<Bson> operations = context.getAggregateOperations();
@@ -699,7 +650,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
         final String dayNameQuery =
                 String.format(
                         "SELECT * FROM \"%s\".\"%s\" WHERE \"field\" <> CURRENT_DATE",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context = queryMapper.get(dayNameQuery);
         Assertions.assertNotNull(context);
         final List<Bson> operations = context.getAggregateOperations();
@@ -728,7 +679,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
         final String dayNameQuery =
                 String.format(
                         "SELECT * FROM \"%s\".\"%s\" WHERE CAST(\"field\" AS TIME) <> CURRENT_TIME",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context = queryMapper.get(dayNameQuery);
         Assertions.assertNotNull(context);
         final List<Bson> operations = context.getAggregateOperations();
@@ -758,7 +709,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
         final String dayNameQuery =
                 String.format(
                         "SELECT * FROM \"%s\".\"%s\" WHERE CAST(\"field\" as TIMESTAMP) <> CURRENT_TIMESTAMP",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context = queryMapper.get(dayNameQuery);
         Assertions.assertNotNull(context);
         final List<Bson> operations = context.getAggregateOperations();
@@ -788,7 +739,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT * FROM \"%s\".\"%s\" " +
                                 "WHERE EXTRACT(YEAR FROM \"field\") = 2021",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context = queryMapper.get(dayNameQuery);
         Assertions.assertNotNull(context);
         final List<Bson> operations = context.getAggregateOperations();
@@ -819,7 +770,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT * FROM \"%s\".\"%s\" " +
                                 "WHERE TIMESTAMPADD(DAY, 3, \"field\") = '2020-01-04'",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context = queryMapper.get(dayNameQuery);
         Assertions.assertNotNull(context);
         final List<Bson> operations = context.getAggregateOperations();
@@ -847,7 +798,7 @@ public class DocumentDbQueryMappingServiceDateTimeTest extends DocumentDbFlapDoo
                 String.format(
                         "SELECT * FROM \"%s\".\"%s\" " +
                                 "WHERE TIMESTAMPDIFF(DAY, \"field\", \"field\") = 0",
-                        DATABASE_NAME, DATE_COLLECTION_NAME);
+                        getDatabaseName(), DATE_COLLECTION_NAME);
         final DocumentDbMqlQueryContext context = queryMapper.get(dayNameQuery);
         Assertions.assertNotNull(context);
         final List<Bson> operations = context.getAggregateOperations();
